# handle incomplete reads in DeflateStream benchmarks